requirements.txt
aiologger==0.6.1
aiounittest==1.4.0
argon2-cffi==20.1.0
asyncpg==0.24.0
asynctest==0.13.0
email_validator==1.1.3
fastapi==0.68.0
fastapi-crudrouter==0.8.0
hypercorn==0.11.2
httpx==0.19.0
inflect==5.3.0
PyYAML==5.4.1
tortoise-orm==0.17.6
python-dotenv==0.19.0
pytz~=2021.1
starlette>=0.14.2
